Initial test of dataframe output
